create generic contract service
